*Update rAthena 70aa7e03ee, 263c138f6d, a8a44254e1, 256e0652b1, 5577161e6e, efdca3e25d, b44e5e11f2, 5e261ba415, 7cd82d00db, 42b29ee81e, 82467e14ed, 267d521c00, 8742f78da98, 32d659fee7, bfac556ecb, 98c6aee157, 8cb97785c2

+++ trunk/src/map/cashshop.h
@@ -43,7 +43,7 @@ enum CASHSHOP_BUY_RESULT {
 };
 
 struct cash_item_data {
-	uint32 nameid;
+	unsigned short nameid;
 	uint32 price;
 };
